+++ addOns/requester/CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Support for renaming tabs.
+- More help.
+
 ### Changed
 - Update minimum ZAP version to 2.11.1.
 - Maintenance changes.
+- Moved Help button to the Response tab.
 
 ## [5] - 2021-10-07
 ### Changed